SisDis/producer_service/requirements.txt
pymongo
pandas
matplotlib
kafka-python